+++ src/constants/tip-text.ts
@@ -77,54 +77,3 @@ export const defaultToolbarTip: Record<string, QuillToolbarTipOptions['tipTextMa
     },
   },
 };
-
-export function createI18nToolbarTipMap(options: {
-  prefix?: string;
-  formats?: {
-    simple?: string[];
-    withValues?: Record<string, string[]>;
-  };
-  i18nModuleName?: string;
-} = {}) {
-  const {
-    prefix = 'toolbar',
-    i18nModuleName = 'i18n',
-  } = options;
-
-  const FORMATS = options.formats || {
-    simple: ['bold', 'italic', 'underline', 'strike', 'color', 'background', 'blockquote', 'code-block', 'code', 'link', 'image', 'video', 'formula', 'clean'],
-    withValues: {
-      list: ['ordered', 'bullet', 'check'],
-      script: ['sub', 'super'],
-      indent: ['-1', '+1'],
-      header: ['', '1', '2', '3', '4', '5', '6'],
-      align: ['', 'center', 'right', 'justify'],
-      size: ['small', '', 'large', 'huge'],
-      font: ['', 'serif', 'monospace'],
-      direction: ['', 'rtl'],
-    },
-  };
-
-  const tipTextMap: QuillToolbarTipOptions['tipTextMap'] = {};
-  // Simple buttons - use onShow to dynamically get translation
-  for (const format of (FORMATS.simple || [])) {
-    tipTextMap[format] = {
-      onShow() {
-        const i18n = this.getModule?.(i18nModuleName) as any;
-        if (!i18n) return;
-        return i18n.t(`${prefix}.${format}`, {});
-      },
-    };
-  }
-  // Buttons with values
-  for (const [format] of Object.entries(FORMATS.withValues || {})) {
-    tipTextMap[format] = {
-      onShow(this: any, target: HTMLElement, value: string) {
-        const i18n = this.getModule?.(i18nModuleName) as any;
-        if (!i18n) return;
-        return i18n.t(`${prefix}.${format}.${value}`, {});
-      },
-    };
-  }
-  return tipTextMap;
-}
