+++ src/supports/index.ts
@@ -1,0 +1,1 @@
+export * from './quill-i18n';
